Move auth stuff out of server.js into own file 	new file:   src/server/config/auth.js 	modified:   src/server/request-handlers/users.js 	modified:   src/server/server.js
